refactor: 상품 가격 Integer -> BigDecimal

+++ product_service/src/main/java/com/cargohub/product_service/application/command/CreateProductCommandV1.java
@@ -1,5 +1,6 @@
 package com.cargohub.product_service.application.command;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreateProductCommandV1(
@@ -7,7 +8,7 @@ public record CreateProductCommandV1(
         UUID firmId,
         UUID hubId,
         Integer stockQuantity,
-        Integer price,
+        BigDecimal price,
         Boolean sellable,
         UUID createdBy
 ) {
