+++ product_service/src/main/java/com/cargohub/product_service/presentation/dto/response/ReadProductDetailResponseV1.java
@@ -2,6 +2,7 @@ package com.cargohub.product_service.presentation.dto.response;
 
 import com.cargohub.product_service.application.dto.ReadProductDetailResultV1;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -9,11 +10,9 @@ public record ReadProductDetailResponseV1(
      UUID id,
      String name,
      UUID firmId,
-//     String firmName,
      UUID hubId,
-//     String hubName,
      Integer stockQuantity,
-     Integer price,
+     BigDecimal price,
      boolean sellable,
      LocalDateTime createdAt,
      UUID createdBy,
